requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
tensorflow==2.21.0
pillow==10.4.0
numpy==2.1.1
python-multipart==0.0.22